+++ resharper/resharper-unity/src/Rider/CodeInsights/UnityCodeInsightFieldUsageProvider.cs
@@ -1,60 +1,53 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using JetBrains.Application.DataContext;
 using JetBrains.Application.Threading;
+using JetBrains.Application.UI.Actions.ActionManager;
+using JetBrains.Application.UI.ActionsRevised.Handlers;
 using JetBrains.Application.UI.Controls;
 using JetBrains.Application.UI.Controls.BulbMenu.Items;
 using JetBrains.Application.UI.Controls.GotoByName;
 using JetBrains.Application.UI.Controls.JetPopupMenu;
+using JetBrains.Application.UI.DataContext;
+using JetBrains.Application.UI.PopupLayout;
 using JetBrains.Application.UI.Tooltips;
 using JetBrains.Diagnostics;
+using JetBrains.DocumentModel;
+using JetBrains.DocumentModel.DataContext;
 using JetBrains.Lifetimes;
 using JetBrains.ProjectModel;
+using JetBrains.ProjectModel.DataContext;
 using JetBrains.ReSharper.Daemon.CodeInsights;
 using JetBrains.ReSharper.Feature.Services.Daemon;
-using JetBrains.ReSharper.Feature.Services.Occurrences;
+using JetBrains.ReSharper.Feature.Services.Navigation.Settings;
+using JetBrains.ReSharper.Host.Features.Services;
+using JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.Navigation.GoToUnityUsages;
+using JetBrains.ReSharper.Plugins.Unity.Feature.Caches;
 using JetBrains.ReSharper.Plugins.Unity.ProjectModel;
-using JetBrains.ReSharper.Plugins.Unity.Resources.Icons;
-using JetBrains.ReSharper.Plugins.Unity.Yaml.Psi;
 using JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.Caches;
-using JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.Caches.UnityEditorPropertyValues;
+using JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.DeferredCaches;
+using JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.DeferredCaches.AssetHierarchy.References;
+using JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.DeferredCaches.AssetInspectorValues;
+using JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.DeferredCaches.AssetInspectorValues.Values;
 using JetBrains.ReSharper.Psi;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
-using JetBrains.ReSharper.Psi.CSharp.Util;
-using JetBrains.ReSharper.Psi.Files;
-using JetBrains.ReSharper.Psi.Modules;
+using JetBrains.ReSharper.Psi.DataContext;
 using JetBrains.ReSharper.Psi.Tree;
 using JetBrains.ReSharper.Psi.Util;
 using JetBrains.ReSharper.Resources.Shell;
 using JetBrains.Rider.Model;
-using JetBrains.TextControl.TextControlsManagement;
 
 namespace JetBrains.ReSharper.Plugins.Unity.Rider.CodeInsights
 {
     [SolutionComponent]
     public class UnityCodeInsightFieldUsageProvider : AbstractUnityCodeInsightProvider
     {
-        private const int POP_UP_MAX_COUNT = 1000;
-        public enum UnityPresentationType
-        {
-            Enum,
-            Bool,
-            String,
-            OtherSimple,
-            FileId,
-            ValueType,
-            Other,
-        }
-
-        private readonly Lifetime myLifetime;
-        private readonly ConnectionTracker myConnectionTracker;
         private readonly UnityApi myUnityApi;
-        private readonly IPsiFiles myFiles;
-        private readonly UnityHost myUnityHost;
-        private readonly UnitySceneDataLocalCache myUnitySceneDataLocalCache;
-        private readonly ITooltipManager myTooltipManager;
-        private readonly TextControlManager myTextControlManager;
-        private readonly UnityEditorProtocol myProtocol;
+        private readonly DeferredCacheController myDeferredCacheController;
+        private readonly AssetInspectorValuesContainer myInspectorValuesContainer;
+        private readonly DataContexts myContexts;
+        private readonly IActionManager myActionManager;
         public override string ProviderId => "Unity serialized field";
         public override string DisplayName => "Unity serialized field";
         public override CodeLensAnchorKind DefaultAnchor => CodeLensAnchorKind.Right;
@@ -62,190 +55,66 @@ namespace JetBrains.ReSharper.Plugins.Unity.Rider.CodeInsights
         public override ICollection<CodeLensRelativeOrdering> RelativeOrderings =>
             new[] {new CodeLensRelativeOrderingLast()};
 
-        public UnityCodeInsightFieldUsageProvider(Lifetime lifetime, UnitySolutionTracker unitySolutionTracker, ConnectionTracker connectionTracker,
-            UnityApi unityApi, UnityHost host, BulbMenuComponent bulbMenu, IPsiFiles files, UnityHost unityHost, UnitySceneDataLocalCache sceneDataCache,
-            ITooltipManager tooltipManager, TextControlManager textControlManager, UnityEditorProtocol protocol)
+        public UnityCodeInsightFieldUsageProvider(UnitySolutionTracker unitySolutionTracker,
+            UnityApi unityApi, UnityHost host, BulbMenuComponent bulbMenu, DeferredCacheController deferredCacheController,
+            AssetInspectorValuesContainer inspectorValuesContainer)
             : base(unitySolutionTracker, host, bulbMenu)
         {
-            myLifetime = lifetime;
-            myConnectionTracker = connectionTracker;
             myUnityApi = unityApi;
-            myFiles = files;
-            myUnityHost = unityHost;
-            myUnitySceneDataLocalCache = sceneDataCache;
-            myTooltipManager = tooltipManager;
-            myTextControlManager = textControlManager;
-            myProtocol = protocol;
+            myDeferredCacheController = deferredCacheController;
+            myInspectorValuesContainer = inspectorValuesContainer;
+            myActionManager = Shell.Instance.GetComponent<IActionManager>();
+            myContexts =  Shell.Instance.GetComponent<DataContexts>();
         }
         
-        private static (string guid, string propertyName)? GetAssetGuidAndPropertyName(ISolution solution, IDeclaredElement declaredElement)
+        private static (string guid, string[] propertyNames) GetAssetGuidAndPropertyName(ISolution solution, IField declaredElement)
         {
             Assertion.Assert(solution.Locks.IsReadAccessAllowed(), "ReadLock required");
             
-            var containingType = (declaredElement as IClrDeclaredElement)?.GetContainingType();
+            var containingType = declaredElement.GetContainingType();
             if (containingType == null)
-                return null;
-
+                return (null, null);
+            
             var sourceFile = declaredElement.GetSourceFiles().FirstOrDefault();
-            if (sourceFile == null)
-                return null;
+            if (sourceFile == null || !sourceFile.GetLocation().NameWithoutExtension.Equals(containingType.ShortName))
+                return (null, null);
 
             if (!sourceFile.IsValid())
-                return null;
+                return (null, null);
 
             var guid = solution.GetComponent<MetaFileGuidCache>().GetAssetGuid(sourceFile);
             if (guid == null)
-                return null;
+                return (null, null);
 
-            // TODO [19.3] support several names for field 
-//            var formerlySerializedAs = AttributeUtil.GetAttribute(fieldDeclaration, KnownTypes.FormerlySerializedAsAttribute);
-//            var oldName = formerlySerializedAs?.Arguments.FirstOrDefault()?.Value?.ConstantValue?.Value as string;
-
-            return (guid, declaredElement.ShortName);
+            return (guid, AssetUtils.GetAllNamesFor(declaredElement).ToArray());
         }
         
         public override void OnClick(CodeInsightsHighlighting highlighting, ISolution solution)
         {
-            if (highlighting is UnityInspectorCodeInsightsHighlighting unityInspectorCodeInsightsHighlighting)
-            {
-                Shell.Instance.GetComponent<JetPopupMenus>().Show(solution.GetLifetime(),
-                    JetPopupMenu.ShowWhen.NoItemsBannerIfNoItems, (lifetime, menu) =>
-                    {
-                        var presentationType = unityInspectorCodeInsightsHighlighting.UnityPresentationType;
-                        var initValue = unityInspectorCodeInsightsHighlighting.ConstantValue;
-                        var declaredElement = (highlighting.DeclaredElement as IClrDeclaredElement).NotNull("declaredElement != null");
+            var rules = new List<IDataRule>();
+            rules.AddRule("Solution", ProjectModelDataConstants.SOLUTION, solution);      
+      
+            var declaredElement = highlighting.DeclaredElement;
+            rules.AddRule("DeclaredElement", PsiDataConstants.DECLARED_ELEMENTS_FROM_ALL_CONTEXTS, new[] {  declaredElement });
 
-                        if (!declaredElement.IsValid())
-                            return;
-                        
-                        var result = GetAssetGuidAndPropertyName(solution, declaredElement);
-                        if (!result.HasValue)
-                            return;
-
-                        
-                        var valuesCache = solution.GetComponent<UnitySceneDataLocalCache>();
-                        var values = valuesCache.GetPropertyValues(result.Value.guid, result.Value.propertyName);
-
-                        menu.Caption.Value = WindowlessControlAutomation.Create("Inspector values");
-                        menu.KeyboardAcceleration.Value = KeyboardAccelerationFlags.QuickSearch;
-
-                        var valuesToShow = values.Where(t => !IsSameWithInitializer(t.Value, presentationType, initValue)).Take(POP_UP_MAX_COUNT);
-                        foreach (var valueWithLocation in valuesToShow)
-                        {
-                            var value = valueWithLocation.Value;
-                            if (ShouldShowUnknownPresentation(presentationType))
-                                menu.ItemKeys.Add(valueWithLocation);
-                            else
-                                menu.ItemKeys.Add(valueWithLocation);
-
-                        }
-
-                        menu.DescribeItem.Advise(lifetime, e =>
-                        {
-                            var value = (e.Key as MonoBehaviourPropertyValueWithLocation).NotNull("value != null");
-
-                            string shortPresentation;
-                            if (ShouldShowUnknownPresentation(presentationType))
-                            {
-                                shortPresentation = "...";
-                            }
-                            else
-                            {
-                                if (!declaredElement.IsValid())
-                                    return;
-
-                                using (CompilationContextCookie.GetExplicitUniversalContextIfNotSet())
-                                {
-                                    var type = declaredElement.Type();
-                                    shortPresentation = GetPresentation(value, presentationType, type, solution,
-                                        declaredElement.Module, true);
-                                }
-                            }
-
-                            e.Descriptor.Text = shortPresentation;
-                            OccurrencePresentationUtil.AppendRelatedFile(e.Descriptor, value.File.DisplayName.Replace("\\", "/"));
-
-                            e.Descriptor.Icon = UnityFileTypeThemedIcons.FileUnity.Id;
-                            e.Descriptor.Style = MenuItemStyle.Enabled;
-                        });
-
-                        menu.ItemClicked.Advise(lifetime, key =>
-                        {
-                            if (!myConnectionTracker.IsConnectionEstablished())
-                            {
-                                ShowNotification();
-                                return;
-                            }
-
-                            var value = (key as MonoBehaviourPropertyValueWithLocation).NotNull("value != null");
-                            
-                            UnityEditorFindUsageResultCreator.CreateRequestAndShow(myProtocol, myUnityHost, myLifetime, solution.SolutionDirectory, myUnitySceneDataLocalCache, 
-                                value.Value.MonoBehaviour, value.File);
-                        });
-                    });
-            }
-        }
-
-        private void ShowNotification()
-        {
-            var textControl = myTextControlManager.LastFocusedTextControl.Value;
-            if (textControl == null)
-                return;
-            
-            myTooltipManager.Show("Start the Unity Editor to view changes in the Inspector", lifetime => textControl.PopupWindowContextFactory.CreatePopupWindowContext(lifetime));
-        }
-
-        private string GetPresentation(MonoBehaviourPropertyValueWithLocation monoBehaviourPropertyValueWithLocation,
-            UnityPresentationType unityPresentationType, IType enumType, ISolution solution, IPsiModule psiModule, bool showOwner)
-        {
-            string baseResult;
-            if (monoBehaviourPropertyValueWithLocation.Value is MonoBehaviourPrimitiveValue)
-            {
-                baseResult = GetRiderPresentation(unityPresentationType, monoBehaviourPropertyValueWithLocation.GetSimplePresentation(solution), enumType, psiModule);
-            }
-            else 
-            {
-                baseResult = monoBehaviourPropertyValueWithLocation.GetSimplePresentation(solution);
-            }
-
-            if (showOwner)
-                baseResult += " in " + monoBehaviourPropertyValueWithLocation.GetOwnerPresentation(solution);
-
-            return baseResult;
-        }
+            using (ReadLockCookie.Create())
+            {               
+                if (!declaredElement.IsValid())
+                    return;
         
-        private string GetRiderPresentation(UnityPresentationType unityPresentationType, string unityValue, IType enumType, IPsiModule psiModule)
-        {
-            if (unityPresentationType == UnityPresentationType.Bool)
-            {
-                if (unityValue.Equals("0"))
-                    return "\"false\"";
-                return "\"true\"";
+                rules.AddRule("DocumentEditorContext", DocumentModelDataConstants.EDITOR_CONTEXT, new DocumentEditorContext(highlighting.Range));
+                rules.AddRule("PopupWindowSourceOverride", UIDataConstants.PopupWindowContextSource,
+                    new PopupWindowContextSource(lt => new RiderEditorOffsetPopupWindowContext(highlighting.Range.StartOffset.Offset)));
+
+                rules.AddRule("DontNavigateImmediatelyToSingleUsage", NavigationSettings.DONT_NAVIGATE_IMMEDIATELY_TO_SINGLE_USAGE, new object());
+      
+                var ctx = myContexts.CreateWithDataRules(Lifetime.Eternal, rules);
+        
+                var def = myActionManager.Defs.GetActionDef<GoToUnityUsagesAction>();
+                def.EvaluateAndExecute(myActionManager, ctx);
             }
-
-            if (unityPresentationType == UnityPresentationType.Enum)
-            {
-                if (!int.TryParse(unityValue, out var result))
-                    return  "...";
-                var @enum = enumType.GetTypeElement() as IEnum;
-                var enumMemberType = @enum?.EnumMembers.FirstOrDefault()?.ConstantValue.Type;
-                if (enumMemberType == null)
-                    return "...";
-                var enumMembers = CSharpEnumUtil.CalculateEnumMembers(new ConstantValue(result, enumMemberType), @enum);
-
-                return string.Join(" | ", enumMembers.Select(t => t.ShortName));
-            }
-
-            return $"\"{unityValue ?? "..." }\"";
         }
-        
-        private bool ShouldShowUnknownPresentation(UnityPresentationType presentationType)
-        {
-            return presentationType == UnityPresentationType.Other ||
-                   presentationType == UnityPresentationType.ValueType;
-        }
-        
-        
+
         public void AddInspectorHighlighting(IHighlightingConsumer consumer, ICSharpDeclaration element,
             IDeclaredElement declaredElement, string baseDisplayName, string baseTooltip, string moreText, IconModel iconModel,
             IEnumerable<BulbMenuItem> items, List<CodeLensEntryExtraActionModel> extraActions)
@@ -256,20 +125,15 @@ namespace JetBrains.ReSharper.Plugins.Unity.Rider.CodeInsights
             var solution = element.GetSolution();
             Assertion.Assert(solution.Locks.IsReadAccessAllowed(), "ReadLock required");
 
-            var result = GetAssetGuidAndPropertyName(solution, declaredElement);
-            if (!result.HasValue)
+            var field = (declaredElement as IField).NotNull();
+            var (guid, propertyNames) = GetAssetGuidAndPropertyName(solution, field);
+            if (guid == null || propertyNames == null || propertyNames.Length == 0)
                 return;
 
-            var guid = result.Value.guid;
-            var propertyName = result.Value.propertyName;
-
-            var cache = solution.GetComponent<UnitySceneDataLocalCache>();
-
-            var field = (declaredElement as IField).NotNull();
             var type = field.Type;
             var presentationType = GetUnityPresentationType(type);
 
-            if (ShouldShowUnknownPresentation(presentationType))
+            if (myDeferredCacheController.IsProcessingFiles() || ShouldShowUnknownPresentation(presentationType))
             {
                 base.AddHighlighting(consumer, element, field, baseDisplayName, baseTooltip, moreText, iconModel, items, extraActions);
                 return;
@@ -279,27 +143,29 @@ namespace JetBrains.ReSharper.Plugins.Unity.Rider.CodeInsights
             var initValue = (initializer as IExpressionInitializer)?.Value?.ConstantValue.Value;
             
             var initValueUnityPresentation = GetUnitySerializedPresentation(presentationType, initValue);
-            var initValueCount = cache.GetValueCount(guid, propertyName, initValueUnityPresentation);
+            var initValueCount = myInspectorValuesContainer.GetValueCount(guid, propertyNames, initValueUnityPresentation);
 
-            if (initValueCount == 0 && cache.GetPropertyUniqueValuesCount(guid, propertyName) == 1) // only modified value
+            if (initValueCount == 0 && myInspectorValuesContainer.GetUniqueValuesCount(guid, propertyNames) == 1) // only modified value
             {
-                var valueWithLocations = cache.GetUniqueValuesWithLocation(guid, propertyName).ToArray(); 
-                Assertion.Assert(valueWithLocations.Length == 1, "valueWithLocations.Length == 1"); //performance assertion
-                displayName = GetPresentation(valueWithLocations[0], presentationType, type, solution,element.GetPsiModule(), false);
-            } else if (initValueCount > 0 && cache.GetPropertyUniqueValuesCount(guid, propertyName) == 2)  
+                var values = myInspectorValuesContainer.GetUniqueValues(guid, propertyNames).ToArray(); 
+                Assertion.Assert(values.Length == 1, "valueWithLocations.Length == 1"); //performance assertion
+                var value = values[0];
+                displayName = value.GetPresentation(solution, field, false);
+            } else if (initValueCount > 0 && myInspectorValuesContainer.GetUniqueValuesCount(guid, propertyNames) == 2)  
             {
                     
                 // original value & only one modified value
-                var values = cache.GetUniqueValuesWithLocation(guid, propertyName).ToArray();
+                var values = myInspectorValuesContainer.GetUniqueValues(guid, propertyNames).ToArray();
                 Assertion.Assert(values.Length == 2, "values.Length == 2"); //performance assertion
-                    
-                var anotherValueWithLocation = values[0].Value.Value.Equals(initValueUnityPresentation) ? values[1] : values[0];
-                displayName = GetPresentation(anotherValueWithLocation, presentationType, type, solution,element.GetPsiModule(), false);
+
+                var anotherValueWithLocation = values.First(t => !t.Equals(initValueUnityPresentation));
+                displayName = anotherValueWithLocation.GetPresentation(solution, field, false);
             }
             
-            if (displayName == null)
+            if (displayName == null || displayName.Equals("..."))
             {
-                var count = cache.GetFilesWithPropertyCount(guid, propertyName) - cache.GetFilesCountWithoutChanges(guid, propertyName, initValueUnityPresentation);
+                var count = myInspectorValuesContainer.GetAffectedFiles(guid, propertyNames) - 
+                            myInspectorValuesContainer.GetAffectedFilesWithSpecificValue(guid, propertyNames, initValueUnityPresentation);
                 if (count == 0)
                 {
                     displayName = "Unchanged";
@@ -310,28 +176,27 @@ namespace JetBrains.ReSharper.Plugins.Unity.Rider.CodeInsights
                     displayName = $"Changed in {count} {word}";
                 }
             }
-
             
-            consumer.AddHighlighting(new UnityInspectorCodeInsightsHighlighting(element.GetNameDocumentRange(),
-                displayName, tooltip, "Property Inspector values", this, 
-                declaredElement, iconModel, presentationType, initValue));
+            consumer.AddHighlighting(new CodeInsightsHighlighting(element.GetNameDocumentRange(),
+                displayName, tooltip, "Property Inspector values", this,
+                declaredElement, iconModel));
         }
         
-        private object GetUnitySerializedPresentation(UnityPresentationType presentationType, object value)
+        private IAssetValue GetUnitySerializedPresentation(UnityPresentationType presentationType, object value)
         {
             if (presentationType == UnityPresentationType.Bool && value is bool b)
-                return b ? "1" : "0";
+                return b ? new AssetSimpleValue("1") : new AssetSimpleValue("0");
 
             if (presentationType == UnityPresentationType.FileId && value == null)
-                return FileID.Null;
+                return new AssetReferenceValue(new LocalReference(0, 0));
 
             if ((presentationType == UnityPresentationType.OtherSimple  || presentationType == UnityPresentationType.Bool) && value == null)
-                return "0";
+                return new AssetSimpleValue("0");
             
             if (value == null)
-                return string.Empty;
+                return new AssetSimpleValue(string.Empty);
 
-            return value.ToString();
+            return new AssetSimpleValue(value.ToString());
         }
         
         private UnityPresentationType GetUnityPresentationType(IType type)
@@ -365,45 +230,21 @@ namespace JetBrains.ReSharper.Plugins.Unity.Rider.CodeInsights
                    myUnityApi.IsDescendantOf(KnownTypes.Component, typeElement);
         }
         
-        private bool IsSameWithInitializer(MonoBehaviourPropertyValue value, UnityPresentationType type, object initValue)
+        private bool ShouldShowUnknownPresentation(UnityPresentationType presentationType)
         {
-            if (value is MonoBehaviourReferenceValue referenceValue)
-            {
-                var fileId = referenceValue.Reference;
-                if (initValue == null && fileId.IsNullReference)
-                    return true;
-                return false;
-            }
-
-            if (value is MonoBehaviourPrimitiveValue primitiveValue)
-            {
-                if (type == UnityPresentationType.Bool)
-                {
-                    if (initValue == null && primitiveValue.PrimitiveValue.Equals("0"))
-                        return true;
-
-                    if (initValue == null)
-                        return false;
-
-                    if (initValue.Equals(true) && primitiveValue.PrimitiveValue.Equals("1"))
-                        return true;
-
-                    if (initValue.Equals(false) && primitiveValue.PrimitiveValue.Equals("0"))
-                        return true;
-                    return false;
-                }
-                
-                if (type == UnityPresentationType.String)
-                {
-                    if (initValue == null && primitiveValue.PrimitiveValue.Equals(String.Empty))
-                        return true;
-                }
-
-                return primitiveValue.PrimitiveValue.Equals(initValue?.ToString());
-            }
-
-            
-            return false;
+            return presentationType == UnityPresentationType.Other ||
+                   presentationType == UnityPresentationType.ValueType;
+        }
+        
+        public enum UnityPresentationType
+        {
+            Enum,
+            Bool,
+            String,
+            OtherSimple,
+            FileId,
+            ValueType,
+            Other,
         }
     }
 }