+++ resharper/resharper-unity/src/Rider/UnityHost.cs
@@ -1,6 +1,10 @@
 ﻿using System;
+using JetBrains.Application.Threading;
+using JetBrains.Application.Threading.Tasks;
+using JetBrains.Lifetimes;
 using JetBrains.ProjectModel;
 using JetBrains.ReSharper.Host.Features;
+using JetBrains.ReSharper.Plugins.Unity.Feature.Caches;
 using JetBrains.Rider.Model;
 
 namespace JetBrains.ReSharper.Plugins.Unity.Rider
@@ -11,13 +15,21 @@ namespace JetBrains.ReSharper.Plugins.Unity.Rider
         private readonly bool myIsInTests;
         private readonly RdUnityModel myModel;
 
-        public UnityHost(ISolution solution, bool isInTests = false)
+        public UnityHost(Lifetime lifetime, ISolution solution, IShellLocks shellLocks, DeferredCacheController deferredCacheController, bool isInTests = false)
         {
             myIsInTests = isInTests;
             if (myIsInTests)
                 return;
 
             myModel = solution.GetProtocolSolution().GetRdUnityModel();
+            deferredCacheController.CompletedOnce.Advise(lifetime, v =>
+            {
+                if (v)
+                {
+                    shellLocks.Tasks.StartNew(lifetime, Scheduling.MainDispatcher,
+                        () => { myModel.IsDeferredCachesCompletedOnce.Value = true; });
+                }
+            });
         }
 
         public void PerformModelAction(Action<RdUnityModel> action)