+++ resharper/resharper-unity/src/Yaml/Psi/Caches/InputManagerAssetHandler.cs
@@ -27,7 +27,7 @@ namespace JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.Caches
         public void Build(IPsiSourceFile sourceFile, ProjectSettingsCacheItem cacheItem)
         {
             {
-                var file = sourceFile.GetDominantPsiFile<UnityYamlLanguage>() as IYamlFile;
+                var file = sourceFile.GetDominantPsiFile<YamlLanguage>() as IYamlFile;
                 var inputs = GetCollection(file, "InputManager", "m_Axes");
 
                 if (inputs == null)