+++ resharper/resharper-unity/src/Yaml/Psi/Caches/MetaFileGuidCache.cs
@@ -3,7 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using JetBrains.Annotations;
 using JetBrains.Collections;
+using JetBrains.DataFlow;
 using JetBrains.Lifetimes;
+using JetBrains.ReSharper.Plugins.Yaml.Psi;
 using JetBrains.ReSharper.Plugins.Yaml.Psi.Tree;
 using JetBrains.ReSharper.Psi;
 using JetBrains.ReSharper.Psi.Caches;
@@ -16,6 +18,8 @@ namespace JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.Caches
     [PsiComponent]
     public class MetaFileGuidCache : SimpleICache<MetaFileCacheItem>
     {
+        private readonly ILogger myLogger;
+
         // We expect to only get one asset with a given guid, but copy/pasting .meta files could break that.
         // CompactOneToListMap is optimised for the typical use case of only one item per key
         private readonly CompactOneToListMap<string, FileSystemPath> myAssetGuidToAssetFilePaths =
@@ -24,10 +28,14 @@ namespace JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.Caches
         // Note that Map is a map of *meta file* to asset guid, NOT asset file!
         private readonly Dictionary<FileSystemPath, string> myAssetFilePathToGuid =
             new Dictionary<FileSystemPath, string>();
+        
+        public Signal<(IPsiSourceFile sourceFile, string oldGuid, string newGuid)> GuidChanged = 
+            new Signal<(IPsiSourceFile sourceFile, string oldGuid, string newGuid)>("GuidChanged");
 
-        public MetaFileGuidCache(Lifetime lifetime, IPersistentIndexManager persistentIndexManager)
+        public MetaFileGuidCache(Lifetime lifetime, IPersistentIndexManager persistentIndexManager, ILogger logger)
             : base(lifetime, persistentIndexManager, MetaFileCacheItem.Marshaller)
         {
+            myLogger = logger;
         }
 
         // This will usually return a single value, but there's always a chance for copy/paste
@@ -50,11 +58,9 @@ namespace JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.Caches
 
         protected override bool IsApplicable(IPsiSourceFile sf)
         {
-            return sf.IsLanguageSupported<UnityYamlLanguage>() && (
-                       sf.Name.EndsWith("cs.meta", StringComparison.InvariantCultureIgnoreCase) ||
-                       sf.Name.EndsWith("unity.meta", StringComparison.InvariantCultureIgnoreCase) ||
-                       sf.Name.EndsWith("prefab.meta", StringComparison.InvariantCultureIgnoreCase)
-                   );
+            return sf.Name.EndsWith("cs.meta", StringComparison.InvariantCultureIgnoreCase) ||
+                   sf.Name.EndsWith("unity.meta", StringComparison.InvariantCultureIgnoreCase) ||
+                   sf.Name.EndsWith("prefab.meta", StringComparison.InvariantCultureIgnoreCase);
         }
 
         public override object Build(IPsiSourceFile sourceFile, bool isStartup)
@@ -62,7 +68,7 @@ namespace JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.Caches
             if (!IsApplicable(sourceFile))
                 return null;
 
-            if (!(sourceFile.GetDominantPsiFile<UnityYamlLanguage>() is IYamlFile yamlFile))
+            if (!(sourceFile.GetDominantPsiFile<YamlLanguage>() is IYamlFile yamlFile))
                 return null;
 
             // Note that this opens the document body chameleon, but we don't care for .meta files. They're lightweight
@@ -86,6 +92,19 @@ namespace JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.Caches
 
         public override void Merge(IPsiSourceFile sourceFile, object builtPart)
         {
+            try
+            {
+                var oldValue = Map.GetValueSafe(sourceFile)?.Guid;
+                var newValue = (builtPart as MetaFileCacheItem)?.Guid;
+                if (!string.Equals(oldValue, newValue))
+                    GuidChanged.Fire((sourceFile, oldValue, newValue));
+            }
+            catch (Exception e)
+            {
+                myLogger.Error(e, "An error occured during changing guid");
+            }
+            
+            
             RemoveFromLocalCache(sourceFile);
             AddToLocalCache(sourceFile, builtPart as MetaFileCacheItem);
             base.Merge(sourceFile, builtPart);
@@ -99,6 +118,16 @@ namespace JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.Caches
 
         public override void Drop(IPsiSourceFile sourceFile)
         {
+            try
+            {
+                var oldValue = Map.GetValueSafe(sourceFile)?.Guid;
+                GuidChanged.Fire((sourceFile, oldValue, null));
+            }
+            catch (Exception e)
+            {
+                myLogger.Error(e, "An error occured during changing guid");
+            }
+            
             RemoveFromLocalCache(sourceFile);
             base.Drop(sourceFile);
         }