+++ rider/src/main/kotlin/com/jetbrains/rider/plugins/unity/actions/ShowUnityLogInRiderAction.kt
@@ -8,12 +8,12 @@ import com.jetbrains.rider.plugins.unity.toolWindow.UnityToolWindowFactory
 class ShowUnityLogInRiderAction : DumbAwareAction() {
     override fun actionPerformed(e: AnActionEvent) {
         val project = e.project?: return
-        e.getHost() ?: return
+        e.getModel() ?: return
         UnityToolWindowFactory.show(project)
     }
 
     override fun update(e: AnActionEvent) {
-        val host = e.getHost()
-        e.presentation.isEnabled = !(host == null || !host.sessionInitialized.valueOrDefault(false))
+        val model = e.getModel()
+        e.presentation.isEnabled = !(model == null || !model.sessionInitialized.valueOrDefault(false))
     }
 }